requirements.txt
chromadb
langchain
langchain-community
langchain-huggingface
langchain-groq
sentence-transformers
python-dotenv
streamlit
